Fix one more type mismatch, xfail 2 tests for GDAL < 2
Resolves #1932

+++ rasterio/_shim1.pyx
@@ -47,8 +47,8 @@ cdef int delete_nodata_value(GDALRasterBandH hBand) except 3:
 
 
 cdef int io_band(
-        GDALRasterBandH band, int mode, float x0, float y0,
-        float width, float height, object data, int resampling=0) except -1:
+        GDALRasterBandH band, int mode, double x0, double y0,
+        double width, double height, object data, int resampling=0) except -1:
     """Read or write a region of data for the band.
 
     Implicit are
